+++ wrapper/frand123.h
@@ -19,7 +19,7 @@ extern "C" {
 // scalar version returns one random number
 double frand123Double_scalar( int64_t *state );
 // vectorial version
-void frand123Double( int64_t *state, const long long lenRes, double *res );
+void frand123Double( int64_t *state, const int64_t lenRes, double *res );
 
 /*
  * generate random single precision numbers uniformly distributed in (0,1)
@@ -33,7 +33,7 @@ void frand123Double( int64_t *state, const long long lenRes, double *res );
 // scalar version returns one random number
 float frand123Single_scalar( int64_t *state );
 // vectorial version
-void frand123Single( int64_t *state, const long long lenRes, float *res );
+void frand123Single( int64_t *state, const int64_t lenRes, float *res );
 
 /*
  * generate random double precision numbers normally distributed with expectation mu and variance sigma
@@ -49,7 +49,7 @@ void frand123Single( int64_t *state, const long long lenRes, float *res );
 // scalar version returns one random number
 double frand123NormDouble_scalar( int64_t *state, const double mu, const double sigma );
 // vectorial version
-void frand123NormDouble( int64_t *state, const double mu, const double sigma, const long long lenRes, double *res );
+void frand123NormDouble( int64_t *state, const double mu, const double sigma, const int64_t lenRes, double *res );
 
 /*
  * generate random single precision numbers normally distributed with expectation mu and variance sigma
@@ -65,7 +65,7 @@ void frand123NormDouble( int64_t *state, const double mu, const double sigma, co
 // scalar version returns one random number
 float frand123NormSingle_scalar( int64_t *state, const float mu, const float sigma );
 //vectorial version
-void frand123NormSingle( int64_t *state, const float mu, const float sigma, const long long lenRes, float *res );
+void frand123NormSingle( int64_t *state, const float mu, const float sigma, const int64_t lenRes, float *res );
 
 /*
  * generate random 64-bit signed integers uniformly distributed over INT64_MIN,..,INT64_MAX
@@ -79,7 +79,7 @@ void frand123NormSingle( int64_t *state, const float mu, const float sigma, cons
 // scalar version returns one random number
 int64_t frand123Integer64_scalar( int64_t *state );
 //vectorial version
-void frand123Integer64( int64_t *state, const long long lenRes, int64_t *res );
+void frand123Integer64( int64_t *state, const int64_t lenRes, int64_t *res );
 
 /*
  * generate random 32-bit signed integers uniformly distributed over INT32_MIN,..,INT32_MAX
@@ -93,7 +93,7 @@ void frand123Integer64( int64_t *state, const long long lenRes, int64_t *res );
 // scalar version returns one random number
 int32_t frand123Integer32_scalar( int64_t *state );
 // vectorial version
-void frand123Integer32( int64_t *state, const long long lenRes, int32_t *res );
+void frand123Integer32( int64_t *state, const int64_t lenRes, int32_t *res );
 
 /*
  * generate random 64-bit unsigned integers uniformly distributed over UINT64_MIN,..,UINT64_MAX
@@ -107,7 +107,7 @@ void frand123Integer32( int64_t *state, const long long lenRes, int32_t *res );
 // scalar version returns one random number
 uint64_t frand123UnsignedInteger64_scalar( int64_t *state );
 //vectorial version
-void frand123UnsignedInteger64( int64_t *state, const long long lenRes, uint64_t *res );
+void frand123UnsignedInteger64( int64_t *state, const int64_t lenRes, uint64_t *res );
 
 /*
  * generate random 32-bit unsigned integers uniformly distributed over UINT32_MIN,..,UINT32_MAX
@@ -121,7 +121,7 @@ void frand123UnsignedInteger64( int64_t *state, const long long lenRes, uint64_t
 // scalar version returns one random number
 uint32_t frand123UnsignedInteger32_scalar( int64_t *state );
 // vectorial version
-void frand123UnsignedInteger32( int64_t *state, const long long lenRes, uint32_t *res );
+void frand123UnsignedInteger32( int64_t *state, const int64_t lenRes, uint32_t *res );
 
 /*
  * initialize the state for the random number generators used (Threefry or ARS)
